Remove preview URL quick open; reorganize project structure
